backend/pyproject.toml
[tool.poetry]
name = "podcast-transcript-api"
version = "0.1.0"
description = "FastAPI backend for Podcast Transcript Search App"
authors = ["Developer"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = ">=0.135.1,<0.136.0"
uvicorn = ">=0.41.0,<0.42.0"
python-dotenv = ">=1.2.2,<2.0.0"
supabase = ">=2.13.0,<3.0.0"
openai = ">=1.59.0,<2.0.0"
pyjwt = "^2.11.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.4"
pytest = ">=9.0.2,<10.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
fixable = ["ALL"]
